test assigness

+++ dao.py
@@ -20,18 +20,18 @@ def open_db(user, password, host, database):
 
 def search(cpf, email):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(
             f"""select * from usuario where cpf = '{cpf}' or email = '{email}'""")
         if has := cursor.fetchall():
             return True
         return False
-    db.close()
+    
 
 
 def save_user(name, sobrenome, email, cpf, telefone, data_nasc, senha):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""insert into usuarios 
         values (default, '{name}','{sobrenome}','{cpf}', '{data_nasc}', '{telefone}','{email}', '{senha}');""")
         db.commit()
@@ -40,18 +40,17 @@ def save_user(name, sobrenome, email, cpf, telefone, data_nasc, senha):
 
 def search_pj(cnpj, email):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(
             f"""select * from usuario where cpf = '{cnpj}' or email = '{email}';""")
         if has := cursor.fetchall():
             return True
         return False
-    db.close()
 
 
 def save_user_pj(name, sobrenome, telefone, email, cnpj, data_nasc, senha):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""insert into usuarios 
         values (default, '{name}','{sobrenome}','{cnpj}', '{data_nasc}', '{telefone}','{email}', '{senha}');""")
         db.commit()
@@ -60,7 +59,7 @@ def save_user_pj(name, sobrenome, telefone, email, cnpj, data_nasc, senha):
 
 def query_exist_email(email):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""select * from usuario where email = '{email}'""")
         if not (list_user := cursor.fetchall()):
             return None, None
@@ -72,10 +71,10 @@ def query_exist_email(email):
 
 def query_exist_cpf(cpf):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""select * from usuario where cpf = '{cpf}'""")
         if not (list_user := cursor.fetchall()):
-            return None, None
+            return None, None, None
         for info in list_user:
             password = info['senha']
             name = info['nome']
@@ -85,10 +84,10 @@ def query_exist_cpf(cpf):
 
 def query_exist_cnpj(cnpj):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""select * from usuario where cnpj = '{cnpj}'""")
         if not (list_user := cursor.fetchall()):
-            return None, None
+            return None, None, None
         for info in list_user:
             password = info['senha']
             name = info['nome']
@@ -99,12 +98,13 @@ def query_exist_cnpj(cnpj):
 def verify_scheduling(start_date, start_hour, final_hour, final_date, id_box):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
 
-    if db:
+    if db and cursor:
         cursor.execute(f"""so elect * from locaca
 where {start_date} >= datainicio and {final_date} <= datafim 
 and {start_hour} >= horainicio or {final_hour} <= horafim
 and id_box = '{id_box}';""")
-    return not cursor.fetchall()
+    if cursor:
+        return not cursor.fetchall()
 
 
 def save_scheduling(start_date, start_hour, final_hour, final_date, id_user, id_box):
@@ -113,14 +113,14 @@ def save_scheduling(start_date, start_hour, final_hour, final_date, id_user, id_
 
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
 
-    if db:
+    if db and cursor:
         cursor.execute(f"""insert into locacao 
         values (default, {start_date},{start_hour},{final_hour},{final_date},{id_user},{id_box});""")
 
 
 def show_all_scheduling():
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute('select * from locacao')
         return cursor.fetchall()
     return False
@@ -128,7 +128,7 @@ def show_all_scheduling():
 
 def show_scheduling_box(id_box):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"select * from locacao where id_box = '{id_box}'")
         return cursor.fetchall()
     return False
@@ -136,7 +136,7 @@ def show_scheduling_box(id_box):
 
 def show_scheduling_user(id_user):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f'select * from locacao where id_user = "{id_user}"')
         return cursor.fetchall()
     return False
@@ -144,7 +144,7 @@ def show_scheduling_user(id_user):
 
 def delete_scheduling(id_scheduling):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""
         delete from locacao where id_locacao = '{id_scheduling}'         
         """)
@@ -154,7 +154,7 @@ def delete_scheduling(id_scheduling):
 
 def update_scheduling(id_scheduling, new_start_date, new_start_hour, new_final_hour, new_final_date):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""
         update locacao set datainicio = '{new_start_date}', horainicio = '{new_start_hour}',
         horafim = '{new_final_hour}', datafim = '{new_final_date}' where id_locaco = '{id_scheduling}';
@@ -165,7 +165,7 @@ def update_scheduling(id_scheduling, new_start_date, new_start_hour, new_final_h
 
 def delete_user(id_user):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""
         delete from usuario where id_user = '{id_user}';               
         """)
@@ -175,7 +175,7 @@ def delete_user(id_user):
 
 def update_user(id_user, new_name, new_lastname, new_cpf, new_birth_date, new_cell, new_email):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""
         update usuario set nome = '{new_name}', sobrenome = '{new_lastname}', cpf = '{new_cpf}',
         data_nascimento = '{new_birth_date}', telefone = '{new_cell}', email = '{new_email}'
@@ -187,7 +187,7 @@ def update_user(id_user, new_name, new_lastname, new_cpf, new_birth_date, new_ce
 
 def new_password(id_user, newpassword):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""update usuario set senha = '{newpassword}' where id_user = '{id_user}';""")
         db.commit()
         db.close()
@@ -195,7 +195,7 @@ def new_password(id_user, newpassword):
     
 def add_box(id_address, name, size, price_hour, description, activated='Y'):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""
         insert into usuario values 
         (default, {id_address}, {name}, {size}, {price_hour}, {description}, {activated})               
@@ -206,7 +206,7 @@ def add_box(id_address, name, size, price_hour, description, activated='Y'):
 
 def ativate_box(id_box):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""
         update box set ativo = 'Y' where id_box = '{id_box}'
         """)
@@ -216,7 +216,7 @@ def ativate_box(id_box):
 
 def desativate_box(id_box):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""
         update box set ativo = 'N' where id_box = '{id_box}'
         """)
@@ -226,7 +226,7 @@ def desativate_box(id_box):
 
 def update_box(id_box, id_address, name, size, price_hour, description, activated='Y'):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""update box 
         set id_endereco = '{id_address}', nome = '{name}', tamanho = '{size}',
         preco_hora = '{price_hour}', descricao = '{description}', ativo = '{activated}'
@@ -237,7 +237,7 @@ def update_box(id_box, id_address, name, size, price_hour, description, activate
 
 def add_address(cep, street, number, complement, district, city, state):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""insert into endereco values (default, '{cep}', '{street}', '{number}',
         '{complement}', '{district}', '{city}', '{state}')""")
         db.commit()
@@ -246,7 +246,7 @@ def add_address(cep, street, number, complement, district, city, state):
 
 def delete_address(id_address):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""
         delete from usuario where id_locacao = '{id_address}';               
         """)
@@ -256,7 +256,7 @@ def delete_address(id_address):
 
 def update_address(id_address, cep, street, number, complement, district, city, state):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""update endereco set cep = '{cep}', rua = '{street}', numero = '{number}',
         complemento = '{complement}', bairro = '{district}', cidade = '{city}', estado = '{state}'
         where id_endereco = '{id_address}';
@@ -267,7 +267,7 @@ def update_address(id_address, cep, street, number, complement, district, city, 
 
 def add_furniture(name):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""insert into tipo_mobilia values (default, '{name}');""")
         db.commit()
         db.close()
@@ -275,7 +275,7 @@ def add_furniture(name):
 
 def delete_furniture(id_fourniture):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""delete from tipo_mobilia where id_mobilia = '{id_fourniture}';""")
         db.commit()
         db.close()
@@ -283,7 +283,7 @@ def delete_furniture(id_fourniture):
 
 def update_furniture(id_fourniture, new_name):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""update tipo_mobilia ser nome = '{new_name}' 
         where id_mobilia = '{id_fourniture}';""")
         db.commit()
@@ -292,7 +292,7 @@ def update_furniture(id_fourniture, new_name):
 
 def add_resource(name):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""insert into tipo_recurso values (default, '{name}');""")
         db.commit()
         db.close()
@@ -301,7 +301,7 @@ def add_resource(name):
 
 def delete_resource(id_resource):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""delete from tipo_recurso where id_recurso = '{id_resource}';""")
         db.commit()
         db.close()
@@ -309,7 +309,7 @@ def delete_resource(id_resource):
 
 def update_resource(id_resource, new_name):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""update tipo_recurso ser nome = '{new_name}' 
         where id_recurso = '{id_resource}';""")
         db.commit()
@@ -318,7 +318,7 @@ def update_resource(id_resource, new_name):
 
 def extra_hour(user_date_scheduling_id, how_many_hours):
     db, cursor = open_db(NAME, PASSWORD, HOST, NAME_DB)
-    if db:
+    if db and cursor:
         cursor.execute(f"""select * from locacao 
         where id_locacao = '{user_date_scheduling_id}';""")
         list_location = cursor.fetchall()
